クラスを分離
